Add custom command to test URLs
Implemented a new custom command that allows users to test a specified URL
